+++ src/NEventStore.Serialization.Gzip.Tests/Properties/AssemblyInfo.cs
@@ -32,4 +32,4 @@ using System.Runtime.InteropServices;
 // [assembly: AssemblyVersion("8.0.0.0")]
 [assembly: AssemblyVersion("8.0.0.0")]
 [assembly: AssemblyFileVersion("8.0.0.0")]
-[assembly: AssemblyInformationalVersion("8.0.0-beta.2+Branch.release-8.0.0.Sha.ee02b4c96b16fd7dddb3464528898ca8917a1ab5")]
+[assembly: AssemblyInformationalVersion("8.0.0-beta.5+Branch.release-8.0.0.Sha.4029904d0dae687763bdf2f11ad4c7754927a8c6")]